Add optional chaining and standardize details cards
