+++ .babelrc
@@ -1,8 +1,3 @@
 {
-  "presets": ["@babel/preset-typescript", "@babel/preset-env"],
-  "plugins": [
-    "@babel/plugin-proposal-class-properties",
-    "@babel/plugin-transform-object-assign",
-    "@babel/plugin-proposal-object-rest-spread"
-  ]
+  "plugins": ["@babel/plugin-proposal-optional-chaining"]
 }
