+++ config-overrides.js
@@ -1,0 +1,2 @@
+const { useBabelRc, override } = require("customize-cra"); // eslint-disable-line
+module.exports = override(useBabelRc());
