+++ package.json
@@ -3,9 +3,9 @@
   "version": "0.8.0",
   "private": true,
   "scripts": {
-    "start": "react-scripts start",
+    "start": "react-app-rewired start",
+    "build": "react-app-rewired build",
     "start:ci": "npm start & wait-on http://localhost:3000",
-    "build": "react-scripts build",
     "lint": "eslint src --ext ts,tsx",
     "format": "eslint --fix src --ext ts,tsx && prettier --write \"**/*.{ts,tsx,json,md}\"",
     "type-check": "tsc --noEmit",
@@ -61,6 +61,7 @@
     "use-query-params": "^0.4.2"
   },
   "devDependencies": {
+    "@babel/plugin-proposal-optional-chaining": "^7.6.0",
     "@babel/preset-typescript": "^7.6.0",
     "@cypress/webpack-preprocessor": "^4.1.0",
     "@percy/cypress": "^2.2.0",
@@ -75,10 +76,11 @@
     "@types/styled-components": "^4.1.19",
     "@types/tti-polyfill": "^0.2.0",
     "@typescript-eslint/eslint-plugin": "^2.3.3",
-    "@typescript-eslint/parser": "^2.3.3",
+    "@typescript-eslint/parser": "^2.6.0",
     "babel-register": "^6.26.0",
     "chalk": "^2.4.2",
     "cross-fetch": "^3.0.4",
+    "customize-cra": "^0.8.0",
     "cypress": "^3.6.0",
     "dotenv": "^8.2.0",
     "eslint-config-prettier": "^6.4.0",
@@ -91,7 +93,8 @@
     "mock-apollo-client": "^0.2.0",
     "node-vibrant": "^3.1.4",
     "npm": "^6.12.1",
-    "prettier": "^1.18.2",
+    "prettier": "^1.19.1",
+    "react-app-rewired": "^2.1.5",
     "react-router-sitemap": "^1.2.0",
     "source-map-explorer": "^2.1.0",
     "ts-loader": "^6.2.1",
