+++ src/components/Card/index.tsx
@@ -1,6 +1,9 @@
 export { default } from "./RawCard";
 export * from "./RawCard";
 
+export { default as DetailsCard } from "./DetailsCard";
+export * from "./DetailsCard";
+
 export { default as CompanyCard } from "./CompanyCard";
 export * from "./CompanyCard";
