+++ src/pages/companies/components/CompanyDetailsCard.tsx
@@ -1,16 +1,15 @@
+/* eslint-disable no-restricted-globals */
+// TODO: upgrade to cra with optional chaining support and remove this stuff
 import React from "react";
 import styled from "styled-components";
 import classNames from "classnames";
 
-import { detailsCardStyles } from "src/theme/snippets";
 import { getLightColor } from "src/shared/utils/color";
 
 import {
-  ISearchFieldProps,
-  Card,
+  IDetailsCardProps,
+  DetailsCard,
   Text,
-  SearchField,
-  Spinner,
   StarRating,
 } from "src/components";
 
@@ -26,19 +25,17 @@ export interface ICompanyDetails {
   color: string;
 }
 
-export interface ICompanyDetailsCardProps extends ISearchFieldProps {
-  loading: boolean;
-  error: boolean;
+export interface ICompanyDetailsCardProps extends IDetailsCardProps {
   companyDetails?: ICompanyDetails;
 }
 
 /*******************************************************************
  *                  **Utility functions/constants**                *
  *******************************************************************/
-const ERROR_OCCURRED_TEXT = "An error occurred while getting company details.";
-
-const getRatingsText = (numRatings: number) => {
+const getRatingsText = (numRatings?: number) => {
   switch (numRatings) {
+    case undefined:
+      return ``;
     case 0:
       return `No reviews yet`;
     case 1:
@@ -48,127 +45,27 @@ const getRatingsText = (numRatings: number) => {
   }
 };
 
-/**
- * Creates the markup for displaying the correct state of
- * the company details, whether still loading, etc.
- * @param loading whether company details are still loading
- * @param error whether fetching company details resulted in error
- * @param details data holding details about the company
- */
-const getDetailsMarkup = (
-  loading: boolean,
-  error: boolean,
-  details?: ICompanyDetails
-) => {
-  if (loading) {
-    return <Spinner className="loading" />;
-  } else if (details) {
-    return (
-      <>
-        <DetailsContainer>
-          <Text variant="heading1" as="div">
-            {details.name}
-          </Text>
-          <Text
-            className="subheading"
-            variant="subheading"
-            color="greyDark"
-            as="div"
-          >
-            {details.desc}
-          </Text>
-          <div className="rating">
-            <StarRating
-              maxStars={5}
-              filledStars={Math.round(details.avgRating)}
-              readOnly
-            >
-              <Text variant="subheading" className="ratingText" color="black">
-                {details.avgRating.toFixed(1)}
-              </Text>
-            </StarRating>
-            <Text
-              variant="subheading"
-              as="div"
-              className="numRatingsText"
-              color="greyDark"
-            >
-              {getRatingsText(details.numRatings)}
-            </Text>
-          </div>
-        </DetailsContainer>
-
-        <Logo src={details.logoSrc} alt={`Logo of ${details.name}`} />
-      </>
-    );
-  }
-
-  // error === true or something has gone horribly wrong
-  return (
-    <Text
-      className="error"
-      variant="subheading"
-      color="error"
-      as="div"
-      align="center"
-    >
-      {ERROR_OCCURRED_TEXT}
-    </Text>
-  );
-};
-
 /*******************************************************************
  *                            **Styles**                           *
  *******************************************************************/
-const Container = styled(Card)`
-  ${detailsCardStyles}
-`;
 
-const MiscContentContainer = styled.div`
+const Container = styled.div`
   display: flex;
   justify-content: space-between;
-  align-items: flex-start;
 
-  position: relative;
-
-  & > .loading {
-    margin: auto;
-    align-self: center;
-    padding: 50px 0;
-  }
-
-  & > .error {
-    margin: auto;
-    align-self: center;
-    padding: 50px 0;
-  }
-`;
-
-const DetailsContainer = styled.div`
-  max-width: 55%;
-
-  & > .subheading {
-    display: inline-block;
-    margin: 10px auto 15px auto;
-  }
-
-  & .rating {
+  & .ratings {
     margin: 35px auto 20px auto;
   }
 
-  & .ratingText {
+  & .rating-text {
     padding: 0 3px;
   }
-
-  ${({ theme }) => theme.mediaQueries.largeMobile`
-    & > .subheading {
-      margin: 0 auto 15px auto;
-    }
-  `}
 `;
 
 const Logo = styled.img`
+  margin-left: 50px;
   max-width: 120px;
+  max-height: 120px;
 `;
 
 /*******************************************************************
@@ -176,32 +73,56 @@ const Logo = styled.img`
  *******************************************************************/
 const CompanyDetailsCard: React.FC<ICompanyDetailsCardProps> = ({
   className,
-  loading,
-  error,
   companyDetails,
-  onTriggerSearch,
-  suggestions,
-  fuseOptions,
+  ...rest
 }) => (
-  <Container
-    className={classNames("company-details-card", className)}
-    color={
-      companyDetails &&
-      companyDetails.color &&
-      getLightColor(companyDetails.color)
-    }
+  <DetailsCard
+    className={classNames("company", className)}
+    color={getLightColor(companyDetails?.color)}
+    {...rest}
   >
-    <MiscContentContainer>
-      {getDetailsMarkup(loading, error, companyDetails)}
-    </MiscContentContainer>
+    <Container>
+      <div>
+        <Text variant="heading1" as="div">
+          {companyDetails?.name}
+        </Text>
+        <Text
+          className="subheading"
+          variant="subheading"
+          color="greyDark"
+          as="div"
+        >
+          {companyDetails?.desc}
+        </Text>
+        <div className="ratings">
+          <StarRating
+            maxStars={5}
+            filledStars={Math.round(
+              companyDetails ? companyDetails.avgRating : 0
+            )}
+            readOnly
+          >
+            <Text variant="body" className="rating-text" color="black">
+              {companyDetails?.avgRating.toFixed(1)}
+            </Text>
+          </StarRating>
+          <Text
+            variant="subheading"
+            as="div"
+            className="num-ratings-text"
+            color="greyDark"
+          >
+            {getRatingsText(companyDetails?.numRatings)}
+          </Text>
+        </div>
+      </div>
 
-    <SearchField
-      onTriggerSearch={onTriggerSearch}
-      suggestions={suggestions}
-      fuseOptions={fuseOptions}
-      inputProps={{ placeholder: "Find a position" }}
-    />
-  </Container>
+      <Logo
+        src={companyDetails?.logoSrc}
+        alt={`Logo of ${companyDetails?.name}`}
+      />
+    </Container>
+  </DetailsCard>
 );
 
 export default CompanyDetailsCard;
