+++ src/pages/jobs/components/JobDetailsCard.tsx
@@ -1,18 +1,17 @@
+/* eslint-disable no-restricted-globals */
+// TODO: upgrade to cra with optional chaining support and remove this stuff
 import React from "react";
 import styled from "styled-components";
 import classNames from "classnames";
 
-import { detailsCardStyles } from "src/theme/snippets";
 import { RouteName } from "src/shared/constants/routing";
 import { getDarkColor } from "src/shared/utils/color";
 
 import {
-  ISearchFieldProps,
-  Card,
+  IDetailsCardProps,
+  DetailsCard,
   Text,
   Link,
-  SearchField,
-  Spinner,
   StarRating,
 } from "src/components";
 
@@ -35,20 +34,17 @@ export interface IJobDetails {
   color: string;
 }
 
-export interface IJobDetailsCardProps extends ISearchFieldProps {
-  loading: boolean;
-  error: boolean;
+export interface IJobDetailsCardProps extends IDetailsCardProps {
   jobDetails?: IJobDetails;
 }
 
 /*******************************************************************
  *                  **Utility functions/constants**                *
  *******************************************************************/
-const ERROR_OCCURRED_TEXT =
-  "An error occurred while getting details for this position.";
-
-const getRatingsText = (numRatings: number) => {
+const getRatingsText = (numRatings?: number) => {
   switch (numRatings) {
+    case undefined:
+      return ``;
     case 0:
       return `No reviews yet`;
     case 1:
@@ -65,183 +61,10 @@ const getRatingsText = (numRatings: number) => {
  * @param error whether fetching job details resulted in error
  * @param details data holding details about the job
  */
-const getDetailsMarkup = (
-  loading: boolean,
-  error: boolean,
-  details?: IJobDetails
-) => {
-  if (loading) {
-    return <Spinner className="loading" />;
-  } else if (details) {
-    return (
-      <DetailsContainer>
-        <Text
-          variant="heading1"
-          as="div"
-          color={details && details.color && getDarkColor(details.color)}
-        >
-          {details.name}
-        </Text>
-        <Link
-          className="subheading"
-          to={`${RouteName.COMPANIES}/${details.companySlug}`}
-          bare
-        >
-          <Text variant="heading3" color="greyDark">
-            {details.companyName}
-          </Text>
-        </Link>
-        {details.location && (
-          <>
-            <span className="separator">&nbsp;{" • "}&nbsp;</span>
-            <Text
-              className="subheading location"
-              variant="heading3"
-              color="greyDark"
-            >
-              {details.location}
-            </Text>
-          </>
-        )}
-
-        <MiscDetails>
-          <div>
-            <StarRating
-              className="rating"
-              maxStars={5}
-              filledStars={Math.round(details.avgRating)}
-              readOnly
-            >
-              <Text variant="subheading" className="ratingText" color="black">
-                {details.avgRating.toFixed(1)}
-              </Text>
-              <Text variant="subheading" className="ratingText" color="black">
-                Overall
-              </Text>
-            </StarRating>
-            <StarRating
-              className="rating"
-              maxStars={5}
-              filledStars={Math.round(details.avgLearningMentorshipRating)}
-              readOnly
-            >
-              <Text variant="body" className="ratingText" color="black">
-                {details.avgLearningMentorshipRating.toFixed(1)}
-              </Text>
-              <Text variant="body" className="ratingText" color="greyDark">
-                Learning &amp; mentorship
-              </Text>
-            </StarRating>
-            <StarRating
-              className="rating"
-              maxStars={5}
-              filledStars={Math.round(details.avgMeaningfulWorkRating)}
-              readOnly
-            >
-              <Text variant="body" className="ratingText" color="black">
-                {details.avgMeaningfulWorkRating.toFixed(1)}
-              </Text>
-              <Text variant="body" className="ratingText" color="greyDark">
-                Meaningful work
-              </Text>
-            </StarRating>
-            <StarRating
-              className="rating"
-              maxStars={5}
-              filledStars={Math.round(details.avgWorkLifeBalanceRating)}
-              readOnly
-            >
-              <Text variant="body" className="ratingText" color="black">
-                {details.avgWorkLifeBalanceRating.toFixed(1)}
-              </Text>
-              <Text variant="body" className="ratingText" color="greyDark">
-                Work-life balance
-              </Text>
-            </StarRating>
-            <Text
-              variant="subheading"
-              as="div"
-              className="numRatingsText"
-              color="greyDark"
-            >
-              {getRatingsText(details.numRatings)}
-            </Text>
-          </div>
-
-          <div className="salary">
-            <Text variant="heading2" as="div">
-              {details.minHourlySalary === details.maxHourlySalary
-                ? details.minHourlySalary
-                : `${details.minHourlySalary} - ${details.maxHourlySalary}`}
-            </Text>
-            <Text variant="heading3" as="div">
-              {`${details.hourlySalaryCurrency}/hr`}
-            </Text>
-          </div>
-        </MiscDetails>
-      </DetailsContainer>
-    );
-  }
-
-  // error === true or something has gone horribly wrong
-  return (
-    <Text
-      variant="subheading"
-      className="error"
-      color="error"
-      as="div"
-      align="center"
-    >
-      {ERROR_OCCURRED_TEXT}
-    </Text>
-  );
-};
 
 /*******************************************************************
  *                            **Styles**                           *
  *******************************************************************/
-const Container = styled(Card)`
-  ${detailsCardStyles}
-`;
-
-const MiscContentContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: flex-start;
-
-  position: relative;
-
-  & > .loading {
-    margin: auto;
-    align-self: center;
-    padding: 50px 0;
-  }
-
-  & > .error {
-    margin: auto;
-    align-self: center;
-    padding: 50px 0;
-  }
-`;
-
-const DetailsContainer = styled.div`
-  width: 100%;
-
-  & > .subheading {
-    display: inline-block;
-  }
-
-  ${({ theme }) => theme.mediaQueries.largeMobile`
-    & .separator {
-      display: none;
-    }
-
-    & .location {
-      display: block;
-    }
-  `}
-`;
-
 const MiscDetails = styled.div`
   display: flex;
   justify-content: space-between;
@@ -252,15 +75,15 @@ const MiscDetails = styled.div`
     justify-content: flex-start;
   }
 
-  & .ratingText {
+  & .rating-text {
     padding: 0 3px;
   }
 
-  & .numRatingsText {
+  & .num-ratings-text {
     margin-top: 4px;
   }
 
-  & > .salary {
+  & .salary {
     display: flex;
     margin-top: auto;
     flex-direction: column;
@@ -271,7 +94,7 @@ const MiscDetails = styled.div`
   ${({ theme }) => theme.mediaQueries.largeMobile`
     flex-direction: column;
 
-    & > .salary {
+    & .salary {
       margin-top: 10px;
       flex-direction: row;
       justify-content: flex-start;
@@ -289,25 +112,110 @@ const MiscDetails = styled.div`
  *******************************************************************/
 const JobDetailsCard: React.FC<IJobDetailsCardProps> = ({
   className,
-  loading,
-  error,
   jobDetails,
-  onTriggerSearch,
-  suggestions,
-  fuseOptions,
+  ...rest
 }) => (
-  <Container className={classNames("job-details-card", className)}>
-    <MiscContentContainer>
-      {getDetailsMarkup(loading, error, jobDetails)}
-    </MiscContentContainer>
+  <DetailsCard className={classNames("job", className)} {...rest}>
+    <Text variant="heading1" as="div" color={getDarkColor(jobDetails?.color)}>
+      {jobDetails?.name}
+    </Text>
+    <Link
+      className="subheading"
+      to={`${RouteName.COMPANIES}/${jobDetails?.companySlug}`}
+      bare
+    >
+      <Text variant="heading3" color="greyDark">
+        {jobDetails?.companyName}
+      </Text>
+    </Link>
+    {jobDetails?.location && (
+      <Text className="subheading location" variant="heading3" color="greyDark">
+        {` • ${jobDetails?.location}`}
+      </Text>
+    )}
 
-    <SearchField
-      onTriggerSearch={onTriggerSearch}
-      suggestions={suggestions}
-      fuseOptions={fuseOptions}
-      inputProps={{ placeholder: "Find a review" }}
-    />
-  </Container>
+    <MiscDetails>
+      <div className="ratings">
+        <StarRating
+          className="rating"
+          maxStars={5}
+          filledStars={Math.round(jobDetails ? jobDetails.avgRating : 0)}
+          readOnly
+        >
+          <Text variant="subheading" className="rating-text" color="black">
+            {jobDetails?.avgRating.toFixed(1)}
+          </Text>
+          <Text variant="subheading" className="rating-text" color="black">
+            Overall
+          </Text>
+        </StarRating>
+        <StarRating
+          className="rating"
+          maxStars={5}
+          filledStars={Math.round(
+            jobDetails ? jobDetails.avgLearningMentorshipRating : 0
+          )}
+          readOnly
+        >
+          <Text variant="body" className="rating-text" color="black">
+            {jobDetails?.avgLearningMentorshipRating.toFixed(1)}
+          </Text>
+          <Text variant="body" className="rating-text" color="greyDark">
+            Learning &amp; mentorship
+          </Text>
+        </StarRating>
+        <StarRating
+          className="rating"
+          maxStars={5}
+          filledStars={Math.round(
+            jobDetails ? jobDetails.avgMeaningfulWorkRating : 0
+          )}
+          readOnly
+        >
+          <Text variant="body" className="rating-text" color="black">
+            {jobDetails?.avgMeaningfulWorkRating.toFixed(1)}
+          </Text>
+          <Text variant="body" className="rating-text" color="greyDark">
+            Meaningful work
+          </Text>
+        </StarRating>
+        <StarRating
+          className="rating"
+          maxStars={5}
+          filledStars={Math.round(
+            jobDetails ? jobDetails.avgWorkLifeBalanceRating : 0
+          )}
+          readOnly
+        >
+          <Text variant="body" className="rating-text" color="black">
+            {jobDetails?.avgWorkLifeBalanceRating.toFixed(1)}
+          </Text>
+          <Text variant="body" className="rating-text" color="greyDark">
+            Work-life balance
+          </Text>
+        </StarRating>
+        <Text
+          variant="subheading"
+          as="div"
+          className="num-ratings-text"
+          color="greyDark"
+        >
+          {getRatingsText(jobDetails?.numRatings)}
+        </Text>
+      </div>
+
+      <div className="salary">
+        <Text variant="heading2" as="div">
+          {jobDetails?.minHourlySalary === jobDetails?.maxHourlySalary
+            ? jobDetails?.minHourlySalary
+            : `${jobDetails?.minHourlySalary} - ${jobDetails?.maxHourlySalary}`}
+        </Text>
+        <Text variant="heading3" as="div">
+          {`${jobDetails?.hourlySalaryCurrency}/hr`}
+        </Text>
+      </div>
+    </MiscDetails>
+  </DetailsCard>
 );
 
 export default React.memo(JobDetailsCard);
